refactor: extract MOTD image-prompt into shared chat module
- Add src/services/chat/motd-image-prompt.ts holding the prompt
  instructions and a provider-agnostic generateMotdImagePrompt
  helper that runs against any OpenAI-compatible client
- OpenAIService and XAIService now delegate to the shared helper,
  removing the duplicated method body
- Rename the param city -> location and reword the prompt so the
  configured location string (city, suburb, or full place name) is
  passed through verbatim
- Document the shared module across the chat/openai/xai/rooivalk
  AGENTS.md files
